requirements.txt
anthropic>=0.19.0
openai>=1.12.0
google-generativeai>=0.3.0
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
backoff>=2.2.0
python-dotenv>=1.0.0